# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/fixture": {
            "get": {
                "description": "Ligin tüm maç fikstürünü döndürür",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "league"
                ],
                "summary": "Tüm maç fikstürünü getirir",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.League"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/league-table": {
            "get": {
                "description": "Mevcut lig tablosunu puan sırasına göre döndürür",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "league"
                ],
                "summary": "Lig tablosunu getirir",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.League"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/play-week": {
            "post": {
                "description": "Ligin güncel haftasını simüle eder ve ligi günceller",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "league"
                ],
                "summary": "Mevcut haftayı oynatır",
                "responses": {
                    "200": {
                        "description": "Week played successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reset-league": {
            "post": {
                "description": "Tüm takımları ve maçları sıfırlayarak ligi yeniden başlatır",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "league"
                ],
                "summary": "Ligi sıfırlar",
                "responses": {
                    "200": {
                        "description": "League reset successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/simulate-all-weeks": {
            "post": {
                "description": "Ligdeki kalan tüm haftaları otomatik olarak simüle eder ve sonuçları döndürür",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "league"
                ],
                "summary": "Tüm ligi simüle eder",
                "responses": {
                    "200": {
                        "description": "Tüm simüle edilmiş maçların sonuçları",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Match"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.League": {
            "type": "object",
            "properties": {
                "current_week": {
                    "type": "integer"
                },
                "matches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Match"
                    }
                },
                "teams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Team"
                    }
                }
            }
        },
        "models.Match": {
            "type": "object",
            "properties": {
                "away_goals": {
                    "type": "integer"
                },
                "away_team_id": {
                    "type": "integer"
                },
                "home_goals": {
                    "type": "integer"
                },
                "home_team_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "played": {
                    "type": "boolean"
                },
                "week": {
                    "type": "integer"
                }
            }
        },
        "models.Team": {
            "type": "object",
            "properties": {
                "draws": {
                    "description": "Yeni eklendi",
                    "type": "integer"
                },
                "goals_against": {
                    "type": "integer"
                },
                "goals_for": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "loses": {
                    "description": "Yeni eklendi",
                    "type": "integer"
                },
                "matches_played": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                },
                "strength": {
                    "type": "integer"
                },
                "wins": {
                    "description": "Yeni eklendi",
                    "type": "integer"
                }
            }
        }
    }
}